pyproject.toml
[project]
name = "daily-tech-digest"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-api-python-client>=2.190.0",
    "google-auth-oauthlib>=1.2.4",
    "notebooklm-mcp-cli>=0.2.18",
]
